+++ .gitignore
@@ -3,7 +3,6 @@ __pycache__/
 *.py[cod]
 *$py.class
 *.out
-scripts/models/**
 
 # C extensions
 *.so
