+++ st_ultralytics/__init__.py
@@ -1,0 +1,5 @@
+__version__ = '0.0.1'
+
+from st_ultralytics.inference import detect_bboxes_from_video
+
+__all__ = '__version__', 'detect_bboxes_from_video'
